requirements.txt
python-telegram-bot
asyncpg
python-dotenv
celery
redis
sendgrid
twilio
pytest
pytest-asyncio